netdog: Update net_config to use test macros
This change updates `net_config` to take advantage of the previously
added test macros.  It removes the majority of the tests from `mod.rs`,
since the test themselves were moved to macros.  It updates the location
of the test data files, and templates the version.  The last and final
change is to add the test macros to `v1`.