buildsys: Use bundle name setting for output
The `bundled-output-path` setting for external files had a small error
when processing the output bundle for Go project dependencies where it
would always use the default value, even when a different value was
provided. This fixes the name of the variable used so it uses the
provided name instead. This already handles setting this to the default
value if the package maintainer has not provided an overridden value.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>