host-ctr: Bump go packages and lock k8s.io/cri-api version
Bumps host-ctr Go dependencies to latest version.

Because containerd is using an older version of the cri-api library from kubernetes
and newer versions of that library carry breaking Go interface changes, this patch
locks `k8s.io/cri-api` to v0.24 which is compatible with containerd's
import. This prevents breaking builds of host-ctr.

Signed-off-by: John McBride <jpmmcb@amazon.com>