netdog: Add `static_lease_path` function to lease module
This change adds an additional public function to the `lease` module
that allows the user to fetch the path for a given interface's static
lease if it exists.  It renames the original `lease_path` function to
`dhcp_lease_path` to be a bit more descriptive, and moves the shared
functionality to a private function.