schnauzer: Add any_enabled template helper
This adds a new template helper called `any_enabled` that can be used in
our templates to check if anything in a collection contains something
with the property of `enabled` set to true.

The helper looks for an array, then checks that the objects within the
array has an "enabled" property. If they do and if any of them are set
to "true" the condition evaluates to be "truey" for conditional blocks
within the template.

If the passed in value is not an array, does not have an "enabled"
property, or has no elements with that property set to true it will
evaluate to falsey.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>