build: refactor image outputs and symlink handling
The larger goal of this change is to make it easier to archive the
contents of a specific build, to allow it to be saved and restored at
a later point in time or on a different machine.

Now that `buildsys` can track and clean up symlinks from builds, move
all symlink generation into the image build process.

Since `buildsys` can also keep track of artifacts in subdirectories,
switch to versioned directories to store the output of image builds.

Previously, "latest" was a directory containing friendly symlinks to
the most recent set of images. Now it is a symlink to the most recent
versioned directory, and the friendly symlinks are created during the
image build.

Signed-off-by: Ben Cressey <bcressey@amazon.com>