CI: extract checks from build actions
We currently have one workflow defined for our GitHub Actions. This
workflow uses a matrix to run tests for every variant defined. In
addition to performing the actual build for these variants, this also
runs unit tests, checks formatting, and checks lint issues. Since these
checks do not change based on the variant, this results in a lot of
duplicated testing and adds a significant delay to how long it takes to
get results.

This change extracts these common checks into a separate workflow so
they are only run once. This also has the benefit of being able to run
in parallel with some of the variant builds and may allow for faster
feedback if there is something like a formatting issue.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>