packages: update libnvidia-container
libnvidia-container added support to configure cgroups through a go
library, which is currently "experimental". In order to use the stable
behavior, this package has to be compiled with `WITH_NVCGO=n`.

This also updates the upstream URL to use a more specific file name
instead of only the version of the package used.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>