Clean up settings information in the README
This sorts the settings values in the README file within each section to
make it a little easier to find what you're looking for.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>

Also adds the content type tags to all code blocks so things like TOML
are formatted and easier to read.

Some settings had example description as a sub-bullet, some had it as
text, and one or two had it outside the bullet list of settings. This
makes it consistent through our README doc.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>