housekeeping: Cancel GH Action on new push
When a PR is proposed to the develop branch, a set of GitHub Action jobs
are kicked off to test the various configuration we need to cover.
With the current configuration of our workflow, if the developer notices
an issue and pushes updates to the source branch it will kick off a new
set of jobs for that updates code, but the jobs running for the previous
revision will keep running.

These jobs take a lot of time and resources to run. Once a new change is
pushed to the source branch, there is no need to continue testing the
older version of the code as it is no longer relevant.

This updates our workflow definition to cancel and currently running
jobs for the same source (`github.ref`) to avoid unnecessary runner load
that may slow down getting results for current PRs.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>