build: fix EFI file system to fit partition size
rpm2img creates the various file systems comprising a Bottlerocket
image, among them the FAT file system that serves as the EFI System
Partition (ESP). rpm2img invokes mkfs.vfat to create the ESP, but
accidentally causes it to size the file system structures for double the
size of the actual ESP.

The mistake is mostly harmless as proven by working images and current
mode of operation that is essentially read-only. However, it can lead to
unexpected problems when trying to write to the ESP. Depending on the
attempted action, failure modes include receiving SIGBUS, receiving EIO,
receiving EINVAL, corrupting files, seeing messages like "lost async
page write" or "attempt to access beyond end of device" in the kernel
log.

The cause of the bug is a somewhat unfortunate interpretation of the
file system size by mkfs.vfat. While its second argument is referred to
as a block count, it is not actually related to the requested sector
size, but always expressed in units of 1 KiB. rpm2img expressed the file
system size in units of sectors (512 bytes), thereby oversizing the file
system by a factor of two.

Signed-off-by: Markus Boehme <markubo@amazon.com>