testsys: Accept `TESTSYS_STARTING_IMAGE_ID`
testsys the cli was able to handle using the `starting_image_id`
argument, but with `cargo make`, the env variable
`TESTSYS_STARTING_COMMIT` is always passed in, so the conflicting args
made `TESTSYS_STARTING_IMAGE_ID` un-usable. This fix removes the
conflicts with, and uses the `TESTSYS_STARTING_IMAGE_ID` whenever it is
provided.