diff-kernel-config: handle absolute paths for output directory
When extracting the kernel config of a just built kernel
diff-kernel-config would always treat the given output directory as a
relative path due to a gratuitous "./" in the path it assembles. This
would try to write to the wrong path when the script was actually
invoked with an absolute path, most likely making it fail because parent
directories wouldn't exist.

Correctly handle absolute paths by dropping the "./". Other parts of the
code dealing with the output directory already don't care whether it's
an absolute or relative path.

Signed-off-by: Markus Boehme <markubo@amazon.com>