prairiedog: treat value-less key differently than "empty" values
Before this change, prairedog did no special handling for value-less
bootconfig keys and would just write out 'key ='. The kernel is
insensitive to whitespace so it would grab the next k/v from the
bootconfig and substitute it for the value.

With this change, prairiedog now differentiates between value-less keys
and keys with empty values (e.g. 'key = ""') both in parsing an existing
bootconfig file as well as when it generates the bootconfig initrd from
existing settings.

When the key is value-less, prariedog will just write out the key
without '=' for its line entry in the bootconfig (e.g. just 'key').