pluto: limit EKS API call timeout to 5 minutes
We don't want pluto to block settings generation for too long during
boot. This limits the retry duration for the EKS describe cluster call
to 5 minutes which matches the IMDS retry duration of 5 minutes
elsewhere. If the API call times out, pluto will fallback to using
default values for the cluster-dns-ip.