netdog: Add `InterfaceId` type to allow for MAC addresses in config
This adds an additional type `InterfaceId`, which allows deserialization
of a valid interface name as well as a valid MAC address from network
configuration.  Validation of the MAC happens during deserialization.

The `WickedInterface` struct has been updated and a `WickedName` struct
has been added to ensure that the "name" node in the serialized XML has
the proper format based on the use of an interface name or MAC address.