netdog: v3: allows MAC address as identifier for interfaces only
This change adds the ability to configure network interfaces using their
MAC address rather than the name udev assigns.  This makes things easier
for situations where MAC is known, but final name is not.  Currently
this ability is limited to interface configuration due to a limitation
with wicked, and may not be used in conjunction with bonds/vlans.