[netdog] Add bonding structs for network configuration
This adds the structures and tests that are used to support bonding in
net.toml. This will be enabled via a version 3 of the net_config module
in a future commit. The current implementation only supports mode 1
(active-backup) for bonding and a handful of settings beyond the minimum
set required for bonding. The code is ready for more modes and settings
to be added when the need arises.