netdog: Handle using MAC address as primary interface identifier
The interface name is important to us in a few places: the wicked
install helper and setting up resolv.conf using DHCP lease.  Previously,
interface names were the sole way to configure interfaces in net.toml.
At network generation time, there isn't a reliable or safe way to map a
MAC to an interface name since udev may still be in the process of
renaming devices.  Later in the boot process (after interfaces have been
setup by wicked and have leases), when interface name is necessary we
can map a MAC address to a name using sysfs.

In order to support MAC addresses as identifiers for interfaces, a few
changes had to be made:

At network generation time, if a MAC address is used for the primary
interface, we write a `primary_mac_address` file rather than a
`primary_interface` file.  The `primary_interface()` function in the
`Interfaces` trait now returns an `InterfaceId` and the
`write_primary_interface()` function determines which file to write
based on the InterfaceId enum variant.

A new function `primary_interface_name()` has been added that returns
the primary interface name.  Under the hood, it reads either the
`primary_interface` or `primary_mac_address` file.  If
`primary_mac_address` is read, it crawls sysfs to determine the network
interface name.

The `write_resolv_conf` subcommand has been updated to use the
`primary_interface_name()` function.