dbus-broker: add patch to disable strict aliasing
The functions in the c-utf8 library have an aliasing rules violation,
and the resulting undefined behavior allows the compiler to ignore
the expected side effects under certain conditions. GCC 11 and beyond
will "optimize" this code in a way that breaks it, if possible.

Normally, the use of `-fPIC` to compile this code stops the compiler
from applying the optimization, since the function is called through
the PLT and could be interposed with some other implementation that
does not violate the rules.

However, we also pass `-fno-semantic-interposition` which tells the
compiler to optimize as though no interposition takes place, allowing
it to exploit the undefined behavior.

Signed-off-by: Ben Cressey <bcressey@amazon.com>