start-local-vm: rework image selection to use latest symlink
The `start-local-vm` hasn't been using the `latest` symlink to find the
set of images to run with yet. That was needlessly complicated and
actually broke with semi-recent changes to buildsys.

Drop the code to clean up old extracted images. All extracted images now
reside in the version-specific directories and can be removed with the
other image build artifacts.

Signed-off-by: Markus Boehme <markubo@amazon.com>