early-boot-config: downgrade warning about inability to iopl to debug
The VMware provider of early-boot-config communicates with the
hypervisor via the VMware backdoor, a couple ports in the x86 port IO
address space. By default Linux does not allow IO port access to user
space processes, and so the used crate optionally invokes the `iopl`
system call to change the IO privilege level of the process. The crate
calls this "privileged" access.

The kernel update in 22a6a58db6c8 ("kernel-5.15: update to 5.15.79")
changed the configuration to disable the `iopl` system call. This causes
the VMware provider to emit a warning about the inability to use the
VMware backdoor via "privileged" access. The provider falls back to
"unprivileged" access, i.e. proceeds to use the backdoor without
changing the IO privilege level first.

This fallback works just as well since both VMware and KVM (in its
VMware guest data emulation [1]) special-case the two IO ports used for
the VMware backdoor. A process trying to access these will only enter
into the hypervisor's fault handler which proceeds to serve the
requests; the hypervisor does not inject a fault into the guest. Hence,
the prior call to `iopl` is redundant, and "unprivileged" access works
just as well.

Only downgrade the warning about the inability to invoke `iopl` to debug
level for now. After some more experience with the behavior on variants
with the 5.15 kernel we can then eliminate the "privileged" access for
the variants with the 5.10 kernel as well.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=9a29d449e3fb7

Signed-off-by: Markus Boehme <markubo@amazon.com>