selinux-policy: adjust default filesystem context
Use "local_t" rather than "any_t" as the default context for xattr
based filesystems.

"unlabeled_t" has been an alias for "local_t" since 7534d22f22684d9,
so this usage is consistent with other policy rules that expect this
as the default label.

The "cni_exec_t" type was previously covered by the rule allowing it
to be associated with ephemeral "any_t" filesystems. With the default
context change, it now needs to be included in the rule for "local_t"
filesystem assocations.

Signed-off-by: Ben Cressey <bcressey@amazon.com>