selinux-policy: widen access to filesystem relabel
The "relabelfrom" and "relabelto" permissions for filesystems are
only used by the SELinux LSM hook for sb_set_mnt_opts() to determine
if the provided "*context=" mount options are allowed by the policy.

This allows containers to mount filesystems using context options,
but only when "local_t" is both the old and new label. This will be
true for the xattr-based filesystems - ext4, xfs, overlayfs - because
of policy defaults, but not otherwise.

The SELinux implementation in the kernel prevents these options from
being changed while the filesystem is active, either by remounting it
or by mounting it again elsewhere. This means that containers cannot
leverage these permissions to interfere with other mounts on the
system; they can only affect newly-mounted filesystems.

The intended use case is to allow containers such as CSI drivers to
"opt out" certain mounts from the filesystem-wide relabel performed
by `containerd` to apply MCS-based isolation. That relabel can be
very costly, if the filesystem has a large number of files. It may
also be blocked by the SELinux policy, in the case of an overlayfs
mounted by a container, since the mounter's credentials do not permit
files to be relabeled.

Signed-off-by: Ben Cressey <bcressey@amazon.com>