packages: Conditionally use `systemd-networkd` based on build flag
This change conditionally builds the `wicked` or `systemd-networkd`
network backend into the variant based on the presence of the
`systemd-networkd` build flag.  This conditional logic is now located in
the os package.  The os package is has been changed to be sensitive to
the `systemd-networkd` image feature so it will be rebuilt if the
feature is added/removed.  `wicked` has been removed from the release
package to prevent it from always being built and included.