sources: box certain sources in error enums
Apply snafu annotations to box the error types most often flagged by
the `clippy::result_large_err` lint, which complains whenever the Err
part of a Result is greater than or equal to 128 bytes.

This is most common for errors with several String fields, since each
String occupies 24 bytes.

That includes:
* apiclient::Error
* datastore::Error
* tough:error::Error
* handlebars::RenderError

Also box a few larger types, like `semver::Version`, and use a Vec
instead of an array for `PartitionSet`.

Signed-off-by: Ben Cressey <bcressey@amazon.com>