migrator: info logging
Before this change, none of migrators logs would make it to the journal.
This made it very hard to understand what was happening when things
went wrong because a problem may have been caused by a previous,
seemingly happy-path iteration of the migration loop.

Now we output a few important info lines to the journal as a record of
what migrator did.