netdog: Enable conditional compilation based on network backend
This commit enables conditional compilation of `netdog` based on the
network backend.  A temporary function has been added that reads the
environment variable `SYSTEMD_NETWORKD`, and emits additional cfgs to
direct compilation.  This environment variable will be passed to the
build if a user has enabled the build flag to use systemd-networkd
rather than wicked in their variant.