release, rpm2img, partyplanner: label BOTTLEROCKET-DATA during firstboot
Changes partitioning to always keep a data partition on the os image.
We grow and label data partition at runtime. We will still create a
separate data partiion on the second volume/disk for "split" partition
layouts.

We add two new services that run as part of 'local-fs.target' and
before 'local.mount' and 'repart-local.service':
  * 'label-data-a.service'
  * 'label-data-b.service'

'label-data-a' and 'label-data-b' will "compete" and both try to
label 'BOTTLEROCKET-DATA' first with the partition they're each waiting
for.

'label-data-a' waits for the data partition that resides on the OS disk
image. Once that device is ready, we call 'systemd-repart' to relabel it
as 'BOTTLEROCKET-DATA' and grow it as much as possible.

'label-data-b' calls 'systemd-repart' to label the data partition on
the data image as 'BOTTLEROCKET-DATA' and grow the partition to fill the
remainder of the disk.

All of this lets the host to boot if data partition on the data image
doesn't exist and the root filesystem disk has leftover extra space to
accommodate a reasonably-sized back-up data partition.