build: drop "BuildArch: noarch" from all packages
This was used as a quick way to disable debuginfo generation, but it
creates ambiguity when using the RPM's architecture to indicate what
target the package was built for.

These packages capture the architecture in the sysroot-based paths
used for installed files, so they are not actually arch-agnostic in
the sense that "noarch" implies.

Signed-off-by: Ben Cressey <bcressey@amazon.com>
(cherry picked from commit 827c051a2ba55e42623d8c50c9bde56eda7c6831)