packages: Update cni and cni-plugin packages
This updates the `cni` and `cni-plugin` packages in Bottlerocket.

Most of the time, the CNI plugins that are used are actually pulled down
via EKS. There are some CNIs, and perhaps metal and vmware variants,
where they need to use the locally installed CNI binaries. This gets
those local binaries up to date with the binaries that end up getting
pulled from EKS.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>