netdog: add primary-interface argument
This argument provides the configured primary interface. This command
provides a way to get the machine's primary interface without reading
the files directly which provides a CLI API to access this value. This
will be used in yet to be implemented logic to watch this interface for
changes on D-Bus.