bloodhound: Add CLI for executing compliance checks
This adds the base bloodhound CLI. The CLI looks for checkers to
execute, runs each one, then reports the results. The results can be
formatted as either a human readable text report, or as JSON output for
programmatic consumption. It can be written directly to a file, or
output to stdout to be viewed in the the shell.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>