selinux-policy: revert change to noatsecure rules
This reverts commit c55908923359554d4ff0bfa2477a21c734de2bee.

The net effect of the previous change was to set the AT_SECURE bit
on execution for all domain transitions for host programs, but not
on executions within containers.

However, `runc` is a host program that needs to transition into the
container domains. Applying AT_SECURE in these cases means that any
environment variables expected to be set for the container's entry
point will be ignored if that program uses `secure_getenv` rather
than `getenv` to access its environment.

The impact is uncertain and depends on how the entry point program is
written and whether it uses `secure_getenv`. In any case, the change
has a larger blast radius than intended; `runc` must be allowed to
use "noatsecure" transitions to maximize compatibility.