check-clippy: Fix user permissions
When running the `check-clippy` task, by default cargo clippy is run
from the SDK. When this container runs, it is under the `builder` user
context. This causes a problem when trying to access the locally mounted
path to the .cargo_lock file. This local file is owned by the current
local user, with only read permissions for Other. This in turn causes a
failure in the clippy execution.

This change updates our check-clippy task to call `docker run` with the
uid:gid of the current user so there are no permission conflicts.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>