Fix aws.profile rendering in cred-provider template
The ecr-credential-provider can use a profile for determining which
credential settings to use out of `~/.aws`. This wasn't getting rendered
in the configuration template because it was inside an `{{#each}}` block
that would change the handlebars context so it would look for settings
values relative to the setting being iterated.

This fixes that by explicitly referencing the `aws.profile` setting from
the root, ignoring iteration context.

Another issue was that setting `aws.profile` after setting the
credential provider settings would not trigger re-rendering of the
credential provider configuration file. This adds an "affected-services"
value for the setting to tell it to restart kubelet, which triggers the
rendering of the template.

A migration was needed because of this affected-services metadata
change.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>