build: fix inference of BUILDSYS_VARIANT-dependent environment variables
A variant string such as "aws-k8s-1.26-nvidia" can be broken down into
its constituent parts, such as platform ("aws"), runtime ("k8s") or
flavor ("nvidia"). Packages are free to adjust their build procedure
based on any of those, though only few make use of this.

When explicitly building an individual package using `cargo make
build-package`, specifying `BUILDSYS_VARIANT` should be enough since all
other information can be deduced from it via the `bottlerocket-variant`
tool, and indeed that's what the script in the Makefile tries to do.
However, it currently fails because of the environment variables
describing the aspects of a variant are unset.

The reason of the failure is subtle and interesting. The
`bottlerocket-variant` tool generates a shell script containing variable
assignments on its stdout, which the Makefile `eval`s:

    eval $(bottlerocket-variant)

The command substitution is unquoted, hence the shell will perform word
splitting on the generated shell script (getting rid of newlines) and
pass words (entire variable assignments such as `FOO=BAR`) as individual
arguments to `eval`. Then `eval` glues its argument back together into
one single source argument (think `$*`) and evaluates it.  Since none of
the variables or values in the generated shell script contain spaces
this would work fine mostly, if it weren't for the fact that for
variants that define image features, `bottlerocket-variant` performs an
`export` of the resulting variable. When `eval`uating the resulting line
of code, all prior variable assignments will be treated as environment
variables only to be set for the execution of the `export` built-in.

To use an example, running `bottlerocket-variant` for `aws-ecs-1` which
currently does not define any image features may generate the following
shell script

    BUILDSYS_VARIANT_PLATFORM=aws
    BUILDSYS_VARIANT_RUNTIME=ecs
    BUILDSYS_VARIANT_FAMILY=aws-ecs
    BUILDSYS_VARIANT_FLAVOR=''

only containing variable assignments. Word splitting would then result
in executing (line break added for commit message)

    eval BUILDSYS_VARIANT_PLATFORM=aws BUILDSYS_VARIANT_RUNTIME=ecs \
         BUILDSYS_VARIANT_FAMILY=aws-ecs BUILDSYS_VARIANT_FLAVOR=''

which appropriately sets all variables. In contrast, the `aws-k8s-1.26`
variant sets an image feature flag and the following shell script is
generated:

    BUILDSYS_VARIANT_PLATFORM=aws
    BUILDSYS_VARIANT_RUNTIME=k8s
    BUILDSYS_VARIANT_FAMILY=aws-k8s
    BUILDSYS_VARIANT_FLAVOR=''
    export BUILDSYS_VARIANT_IMAGE_FEATURE_GRUB_SET_PRIVATE_VAR=1

After word splitting, `eval` is thus invoked (line breaks added for
commit message)

    eval BUILDSYS_VARIANT_PLATFORM=aws \
         BUILDSYS_VARIANT_RUNTIME=k8s \
         BUILDSYS_VARIANT_FAMILY=aws-k8s \
         BUILDSYS_VARIANT_FLAVOR='' \
         export BUILDSYS_VARIANT_IMAGE_FEATURE_GRUB_SET_PRIVATE_VAR=1

Since `export` is a built-in, all variable definitions preceding it are
interpreted to be environment variables only to be valid in the scope of
executing `export`, and the only variable assignment outliving `eval` is
the one for the image feature flag. A subsequent check for the existence
of the other variables consequently fails and the build aborts.

Fix this by properly quoting the command substitution in the Makefile.
This prevents the shell from word splitting and retains all newlines,
thereby passing only one argument to `eval` and forcing it to process
all variable assignments independently.

Bonus section: The preceding explanation only talks about the script for
`build-package` while the same change is performed on the script for
`build-variant`. That's because only `build-package` actually is in need
of the change! `build-variant` is changed for consistency.

The behavior between the tasks is different because `build-package` uses
bash to interpret the shell script. By default, bash deviates from
behavior specified by POSIX in several places. It can be made to better
comply with POSIX in its POSIX mode. The docs at [1] contain a list of
differences to the bash defaults that are activated in POSIX mode. The
change of relevance to this situation is:

    34. Assignment statements preceding POSIX special builtins persist
	in the shell environment after the builtin completes.

POSIX mode can be entered in various ways, e.g. by passing `--posix` to
bash or by invoking it with `sh` in `argv[0]` (achieved e.g. by invoking
it via an `sh` symlink). The latter is how even on systems where bash is
used as implementation of `sh` (e.g. on Fedora, vs. dash on Debian and
Ubuntu) the script for `build-variant` is unaffected by all of the
above.

[1]: https://www.gnu.org/software/bash/manual/html_node/Bash-POSIX-Mode.html

Signed-off-by: Markus Boehme <markubo@amazon.com>