release: ensure data partition can be grown to 20 GiB
This ensures the data partition can be grown to 20GiB - (GPT header and
footer) when we're labeling/growing the partition. If there isn't enough
disk space, systemd-repart fails and skips labeling the partition. This
should help in situations where there's extra space on the root disk for
a data partition and a separate data disk is also present. If the data
partition is created on the root disk, then it will at least have 20GiB
for /local fs storage.