netdog: Add write-primary-interface-status for networkd
wicked provides a hook when interfaces are configured to run extra
actions. These actions are used to write additional configuration needed
by other services and tools in the system. systemd-networkd does not
have the same hook so new functionality is needed to accomplish the same
tasks. For now, the code is parallel to the wicked-specific install.rs
but adds a new cli argument write-primary-interface-status to netdog.
This command is intended to be called when the primary network interface
moves to configured, at that point networkctl status is called for that
configured interface to fetch the required info and IP Address and DNS
configuration is written to the desired files like /etc/resolv.conf.

This cli argument is only built in if the variant has systemd-networkd
as a backend. It will be called by a yet to be built daemon that will
watch for these configured events on D-Bus and replicate the same
functionality wicked currently provides with its install hooks.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>