release: create data filesystem in local-fs.target
The local filesystem is expected to be ready before local-fs.target
is reached.

Signed-off-by: Ben Cressey <bcressey@amazon.com>

We don't need to wait for the 'repart' units since this cause boot to
hang until the repart units timeout on waiting for a potentiaily
non-existent data partition.

'systemd-repart' and 'systemd-makefs' both lock on the block device
before operating on it, so repart is guaranteed to finish before makefs
can create the filesystem.

Co-authored-by: Erikson Tung <etung@amazon.com>