Makefile: add check-migrations task to check for common problems
Add a `check-migrations` task that runs as part of `check`. It attempts
to check for common problems with migrations. It might be nice to have
it run during build tests. While it does not actually run migrations or
check their logic, it can spot the kind of inconsistencies that are
easily lost in reviews.

Signed-off-by: Markus Boehme <markubo@amazon.com>