kernel-5.15: Default disable panic on hung task and lockups
Setting panic as defaults for hung tasks and lockups is probably not the
best idea for the general case. In scenarios with remote storage we may
see hangs as interactions with that storage might be delayed due to
network connectivity issues. In such cases there is use cases that do
not benefit from panic and reboot.

Disable the default for those panic options. This does not take away the
option to configure panic on hung tasks or lockups through either sysctl
or kernel boot parameters.

As a bonus this aligns both kernels 5.10 and 5.15 on coherent settings
for the panic choices.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>