Actions: Only use GOPROXY=direct on nightly build
This adds a new workflow to run nightly that will use the existing
GOPROXY=direct setting to verify that the caching of Go modules is not
hiding an issue with upstream module availability changes.

It trims down the normal PR workflow to skip the `-dev` variants and not
perform the additional checks like unit-tests and shellchecks.

With the addition of the nightly workflow to identify these issues, this
removes the GOPROXY setting from the `build` workflow that runs on each
PR. This should reduce the number of failed jobs that need to be rerun
due to throttling of go module fetches that seems to happen very often
during normal working hours.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>