grub: move embedded public key into dedicated section
For the EFI build, the Secure Boot config signing key is embedded into
the GRUB image so that GRUB can verify the signature of any files it
loads, e.g. its configuration. By default, the public key is embedded
into a section alongside various executable GRUB modules. This set of
patches instead moves the key into its own dedicated section in the
image, where it can easily be replaced with tools such as objcopy or dd
to reuse a set of binary artifacts with different Secure Boot keys.

Signed-off-by: Markus Boehme <markubo@amazon.com>