build: sign artifacts for Secure Boot
When Secure Boot is enabled for the variant, ensure that EFI binaries
and the GRUB configuration are signed with the expected keys, and
replace the placeholder certificates and keys with the right ones.
When building an OVA, populate the template with the EFI variable
data.

Because GRUB will have an embedded GPG public key in this mode, it
will automatically verify the detached signature for every file read.
This is the desired behavior for `grub.cfg`, which contains sensitive
parameters like the dm-verity root hash.

However, it's redundant with the EFI binary verification performed by
the shim verifier. It would also prevent reading the "initrd" that
contains the Boot Config data for kernel command line parameters,
because this is generated on the local system and cannot be signed by
a trusted key.

Since this is the only remaining file to read that is not an EFI
binary, it's OK to disable signature checking inside the verified
`grub.cfg`.

Have GRUB reboot if the kernel cannot be loaded, and backstop this by
blocking edits to menu entries with an empty superusers group. This
prevents runtime modifications to the expected configuration, which
could otherwise be used to alter the dm-verity root hash.

Signed-off-by: Ben Cressey <bcressey@amazon.com>
Signed-off-by: Markus Boehme <markubo@amazon.com>