grub: add SBAT info
shim expects to find an SBAT section in GRUB, and will not continue
booting if it is missing.

Bottlerocket's build of GRUB is downstream from Amazon Linux, which
is downstream from Fedora and RHEL. However, most of GRUB's modules
are not included in the BIOS and EFI images, and the configuration
file built into the image disables loading any other modules. Hence
it's not a given that vulnerabilities in either upstream should lead
to SBAT-based revocations.

Bottlerocket carries quite a few out-of-tree patches which add the
`gpt` and `gptprio` modules, and consequently has an unenviable, but
useful, claim to its own vendor entry. Clearly the vast majority of
GRUB development happens elsewhere, and the use of a distro-specific
vendor entry is not meant to imply otherwise.

There are no current plans for Bottlerocket to participate in the
wider ecosystem of Secure Boot for Linux distributions, by way of a
Microsoft-signed shim, so the choice of SBAT metadata is not relevant
elsewhere.

Signed-off-by: Ben Cressey <bcressey@amazon.com>