shim: embed placeholder for vendor cert
Add a large placeholder certificate to the shim binary to facilitate
replacement in the final stage of the build.

Note that shim is used as the first-stage UEFI bootloader for all
variants, but only enforces Secure Boot if the relevant EFI variables
are set, and will not use the vendor certificate otherwise.

Signed-off-by: Ben Cressey <bcressey@amazon.com>