build: add scripts to create Secure Boot profiles
For Secure Boot, various certificates, keys, and configuration files
are needed to sign binaries and register images. Provide two scripts
to simplify the process of generating the correct artifacts.

The "local" version of the script is only meant for non-production
use, for example by individual developers or by automated CI testing,
where the variant builds must support the feature but do not need to
be maintained indefinitely.

The "aws" version of the script expects various AWS resources such as
private CAs and managed keys to be available. This can be costly and
only makes sense for official builds of supported variants.

Signed-off-by: Ben Cressey <bcressey@amazon.com>