variants: enable Secure Boot for dev builds
Secure Boot can't be added in a backwards-compatible way to existing
variants without breaking the downgrade path, but will be the default
for all newer variants.

To ensure the functionality is exercised, set the Secure Boot feature
flag in the *-dev variants, which are not intended for production use
and do not support downgrades.

Signed-off-by: Ben Cressey <bcressey@amazon.com>