kernel-5.10: increase default net.unix.max_dgram_qlen to 512
Increase the kernel's default value for the net.unix.max_dgram_qlen sysctl
to 512. This is a change to the kernel rather than a plain sysctl setting
since systemd-sysctl only applies settings to the host, while the changed
default value in the kernel also applies to every new network namespace.

Signed-off-by: Markus Boehme <markubo@amazon.com>