nvidia-k8s-device-plugin: change service dependency
The NVIDIA device plugin knows how to re-connect to the kubelet in case
it isn't ready or it re-starts. Thus, it is sufficient to declare the
dependency on the kubelet with "Wants" instead of a stronger dependency
directive.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>
(cherry picked from commit c63e3b3978d9139f7fddc0895ea8cd8f43cb9a54)