build: clean up boilerplate
Previously we used empty `pkg.rs` and `lib.rs` files to carry out a
no-op build. Since `rustc` is equally happy compiling an empty file
in another directory, we can use `/dev/null` instead.

Likewise, we used a copy of the same build script in all variants and
almost all packages. Instead, we can have a single copy of the script
in the parent directories and refer to it via relative path.

Signed-off-by: Ben Cressey <bcressey@amazon.com>