wicked: reduce initial DHCPv6 solicitation delay to 100 ms
Reduce the maximum initial solicitation delay to 100 ms from the
spec-mandated value of 1000 ms to reduce the average and variance in
time to node readiness.

This is not strictly compliant with RFC 3315 (treating the initial delay
as a MUST), and only tolerated by its successor RFC 8415 (changing it to
a SHOULD). This change therefore is not a good candidate for upstreaming.

Signed-off-by: Markus Boehme <markubo@amazon.com>