netdog: Add wicked IPv6 config structures to support RA
This change adds the necessary structures to wicked in order to support
setting the `accept_ra` IPv6 option via interface config.  Doing this
in config allows wicked to manage this option and avoids any race
conditions that may occur between wicked and the kernel when attempting
to set this via sysctl.