open-vm-tools: Modify shutdown for Bottlerocket
Bottlerocket doesn't have /bin/sh which makes the call system() fail
with an unexpected error code. This results in the guest shutdown and
reboot calls via vmtoolsd returning as if the request was successful but
the call actually failed completely. This commit replaces the current
logic for executing a shutdown|reboot with g_spawn_command_line_sync()
which should be a drop in replacement but removes the /bin/sh
requirement.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>