docs: Remove link for security feature roadmap
The SECURITY_FEATURES.md file contained a link to the Bottlerocket
project roadmap, filtering on security items. This category is not
currently used in the project roadmap, so the link from this page would
lead to an empty project board.

This removes the hyperlink since it is no longer useful and tweaks a
couple smaller items noticed while reading through the doc.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>