netdog: Add stronger typing to bond config
Previously bond config implemented validation of the `kind` parameter
using a custom `Deserialize` function, which is made use of by the
`[serde(remote = "Self")]` annotation.  This annotation generates
another implementation of Deserialize, which in this case is unnecessary
as we can do validation via stronger type.

This commit ads a single variant enum `BondKind` which adds the same
validation previously contained in the `Deserialize` function.  The
enum is meant solely to validate the `kind` is equal to "bond" or
"Bond".