netdog: Add stronger typing to vlan config
Previously, vlan config implemented validation in a custom `Deserialize`
function, which is made use of by the `[serde(remote = "Self")]`
annotation.  This annotation generates another implementation of
Deserialize, which in this case is unnecessary as we can do validation
via stronger types.

This change adds a new type wrapper type `VlanId` for VLAN ID.
Previously we stored the ID as a u16 and did the validation via the
deserialize function.  Since ID will be used elsewhere in the upcoming
networkd changes, it makes sense to make it a stronger type.  This type
validates the correct range by implementing `Deserialize` directly.

Along with the addition of the `VlanId` type, this commit adds a single
variant enum `VlanKind`.  This enum is meant solely to validate the
`kind` is equal to "vlan", "VLAN", or "Vlan".