docs: correct dm-verity text in security features
Partial revert of b7c1ee45071202a87436203c8e44d4fa2eb94420.

"Fail close to when the block device is modified" is not correct; in
the absence of memory pressure to evict previously read blocks from
the cache, the detection could occur after much time has passed.

"Fail closed" is correct in that the kernel will panic and reboot if
corruption is detected.

Add a short explanation of the purpose and the value of this design
choice.

Signed-off-by: Ben Cressey <bcressey@amazon.com>