Settings: log set operations
This adds info level logging to capture some trace of setting updates in
the system log at the default logging level. This is to help capture a
record of system changes being made.

This is only a baby step towards getting better system modification
auditing in place, but it is a very minor change that can be done now
until a more comprehensive plan can be put in place.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>