pubsys: Make AMI already registered message more visible
Pubsys emits a message when asked to publish an AMI that is already
found. This is currently a single line message that blends in with the
other output when doing `cargo make ami` and makes it very easy to miss.

Normally this is good, but when actively developing and making local
code changes, this can make it more likely that the developer will not
notice the message and spend time trying to understand why their code
changes are not visible when running new instances using this AMI.

This changes the warning message to be a little more visible so it is
much more noticable when trying to publish an AMI that is already
registered to try to avoid these cases.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>