tools/diff-kernel-config: Add resume functionality
Sometimes, the script may fail for a specific variant after we already
have spent time successfully building other variants. In order to not
let that time go to waste, enable resuming of previous runs of the
script. Be aware that the invocation has to be the same as the original
invocation, plus the resume flag.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>