pluto: increase retries for requests out to AWS
This adds more resiliency when pluto makes API requests out to EC2 and
EKS. We choose a large number of retries to force the SDK to keep
retrying until we timeout on the overall request.

It's now less likely for the request to overall fail due to
transient connection errors.