systemd-networkd: Conditionalize hostnamed/timezoned DBUS calls
systemd-networkd registers a function to call when first connecting to
DBUS.  This function makes three calls to other DBUS services (hostnamed
and timedated) which aren't used in Bottlerocket.  Calls to the same
DBUS services are also made in the DHCP clients.  This change makes a
patch to systemd-networkd that conditionalizes these calls based on the
underlying service being built, return 0 if the services isn't built.

Removing the calls to non-existent services cleans up some confusing and
inconsequential messages in the journal on boot.