kernel-6.1: Make objtool build dependent on CONFIG_OBJTOOL
In the past we have added objtool as a dependency to build external
kmods. However, objtool is dependent on architectural features and other
config settings and is thus not available on all architectures. For
Bottlerocket this means that objtool is not available on arm64. Since
5.15 a config option CONFIG_OBJTOOL has been introduced which we can use
decide if objtool is supported in the current build scenario.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>