netdog: Add top-level interface for networkd config creation
This change adds a new type `NetworkDConfig`, which is the "top-level"
interface for systemd-networkd config creation.  The type can be created
from any version of network config, provided the struct implements the
`TryInto<NetworkDDevice>` trait.  The type presents a single method
`create_files()` which will drive the underlying devices to create and
return a list of appropriate configuration files.