logdog: Add filters for newer sensitive settings
Logdog has a mechanism in place for filtering settings from the
collected output for sensitive data if the setting matches on of the
patterns to exclude. There have been a few settings added since its
introduction that could potentially contain passwords or other senitive
information. This adds current setting names that were not currently
covered by the existing patterns.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>