netdog: Add ability to generate systemd-networkd config
This commit puts all the pieces together to enable systemd-networkd
config generation via netdog.

The change adds an additional trait method to the Interfaces trait that
all versions of net config must implement: `as_networkd_config()`.  This
method returns a `NetworkDConfig`, from which all config files may be
generated.  Under the hood, this method takes advantage of the
TryFrom conversions also included in this commit.

This commit adds the trait method to all versions of net config, and
also adds integration tests to ensure we generate the expected files for
each device in the proper format.  These tests use the same
`net_config.toml` file as the wicked integration tests.