actions: Set nightly continue-on-failure to false
With this set to true, even though steps were failing in the nightly
runs, the overall workflow status was showing green.

There are no other steps that need to run, so no reason to continue on
failure for this job. If any of the jobs fail, we want that failure
status to reflect in the overall run status.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>