bloodhound: Skip K8s 4.2.10 as not applicable
This changes the 4.2.10 check to be a manual test response. This check
does not apply to EKS or EKS-D since external IAM auth is used.

This leaves the validation logic in place as we may want to use it in
the future. For not EKS-related use cases, it may make sense to enable
configuration of certificate rotation. If/when that is done, this check
can be reenabled and updated to validate under the proper conditions.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>