whippet: add whippet to sources
Whippet is a D-Bus listener that subscribes to the network1 D-Bus
service on the system bus. This listens for events that happen to the
primary interface and call netdog to notify it so netdog can update
anything that has changed with the primary interface configuration.

This commit adds a rudimentary but functional version of whippet. There
is more work to be done to make it a robust, reliable daemon.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>