variants: metal: Add linux-firmware package
With the metal-* variants we are targetting a wider variety of hardware
platforms. In order to enable users to properly use Bottlerocket on
platforms we provide drivers for we also need to supply firmware
binaries for some of these devices. Ship these in the metal variants
through new package linux-firmware.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>