netdog: Add VlanLink type parameter for NetworkBuilder
This commit adds a new type parameter for the NetworkBuilder type
specifically for interfaces meant only to be used as the link for a
VLAN.  These types of interfaces are typically used for "tagged
only"-type setups.  The interface requires config so it can be managed
and a member of a VLAN but otherwise has all DHCP and addressing turned
off.