Include systemd-logind for inhibit functionality
Settings were added for kubelet to be configured with pod shutdown grace
policy settings. This feature allows kubelet to hold off a system
shutdown until it has had a chance to terminate and migrate pod
workloads cleanly.

Unfortunately, it was missed when these settings were added that the
mechanism kubelet uses internally for this is to use systemd-inhibit
that is part of systemd-logind. Without the systemd-logind functionality
included in Bottlerocket, the attempt to start processes with inhibit
protection would fail and shutdown calls would be immediate.

This adds systemd-logind to the systemd build process to enable this
functionality, excluding things like PAM and its CLI since those are not
needed for our inhibit requirements.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>