testsys: Fix support for metal migration testing
Previously all migration testing required a bottlerocket crd to be
present, but since EKS Anywhere is used to provision machines, the
bottlerocket crd is not created for metal. This means the migration
agent needs to use the bottlerocket crd for instance ids if it's
available and if not, rely on the cluster crd to get the instance ids.