Only limit concurrency while building RPMs
cargo make build-variant and build-package typically run several RPM
builds concurrently, and those allow for their own levels of
parallelism. Leaving this unrestricted would very easily lead to
context switching overhead.

On the other hand, the rest of the build commands are mostly focused
on building a directory of rust sources. It shouldn't be a problem to
let cargo use as many cores as it needs for those.