systemd: tidy macros and excludes
Switch to standard macros for sysusers.d, tmpfiles.d, and the systemd
unit directory. Although functionally equivalent, the macros help to
communicate "what" is being packaged rather than just "where" it is
on the filesystem. Cleaning up the non-macro paths improves the odds
that a future packager will use the macros as intended.

Reorganize the excluded files to distinguish the ones that are never
included from the ones that are included in subpackages.

Signed-off-by: Ben Cressey <bcressey@amazon.com>