release: consult resolved for host lookups
This commit adds `resolve` to the nsswitch.conf.  It is placed second in
line for hosts resolution because Bottlerocket manages the hosts file in
the case there are custom settings.  `resolve` is annotated with
`[!UNAVAIL=return]` which means the library will move on to the the next
service in the list if it is unavailable, like when using wicked.  The
systemd-resolved package includes libnss_resolve.so, so resolved will be
consulted in that case.

Co-authored-by: Ben Cressey <bcressey@amazon.com>
Co-authored-by: Zac Mrowicki <mrowicki@amazon.com>