xfsprogs: add xfs_info cli
The cli can be used as:
- xfs_info -V
- xfs_info $target where target is mount_point or block_device

This works for both mounted and unmounted target. It fetches data using
xfs_spaceman when the target is mounted and use xfs_db in case of
unmounted target.

Param "-t" for mtab is not included in this cli as
xfs_spaceman -p xfs_info -c "info" -t "$mtab" "${target}"
do not produce expected output.