os/netdog: Create /etc/systemd/network in netdog, remove mount unit
This commit removes the mount unit for /etc/systemd/network in favor of
creating the directory in netdog.  A mount unit was previously used for
this in order to set a specific SELinux context and avoid any races with
tmpfiles.  It was found that the context isn't needed since the programs
that write to it have the required permissions, and netdog itself can
write the directory, avoiding any races.