netdog: Add write-network-status unit for networkd/resolved
The wicked network backend provides a user-configurable hook that can be
used to do things when bringing up an interface.  systemd-networkd does
not provide such a hook.  This change adds a new systemd unit meant to
be used in the context of a system using systemd-networkd as the network
backend.  The unit is meant to write the same sort of information as the
wicked "install" hook; information like the current ip and symlink to
the DNS resolv.conf that is required by units or programs that run later
in boot.  The unit runs early in boot, directly after the
network-online.target but before services that require information
written as part of this hook.

This change updates the previously added
"write-primary-interface-status" subcommand, which is not currently in
use but meant for this purpose.  It has been updated to reflect the
current DNS strategy (systemd-resolved) and does not directly write a
resolv.conf as it did previously.