kernel-6.1: cherry-pick fix for CVE-2023-20588 ("DIV0")
The fix for CVE-2023-20588 is currently only available in the knerel.org
upstream 6.1 stable kernel, but not yet in an Amazon Linux kernel
release. Cherry-pick it from the upstream kernel. Bring in an extra
commit (x86/bugs: Increase the x86 bugs vector size to two u32s) as a
dependency.

Contextual changes are necessary to make these patches apply as we
currently carry Amazon Linux' patches mitigating GDS and SRSO instead of
upstream vairants of these patches.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>