microcode: Add amd ucode mitigating AMD Inception
The AMD ucode to mitigate the Inception vulnerability on Zen4 AMD cpus has
been added to the upstream linux-firmware repository a few days after the
last regular release. Pick this up in order to ship the fixed microcode
before the we get a chance for the next regular linux-firmware release.

This patch can be reverted once we update to an upstream version that
includes it already (most likely the September 2023 release).

Signed-off-by: Leonard Foerster <foersleo@amazon.com>