early-boot-config: read additional local files
This adds two new sources of user-data settings for all platforms:
* `/local/user-data-defaults.toml`
* `/local/user-data-overrides.toml`

The defaults are applied before any other source of user-data, while
the overrides are applied after. This allows sites to customize the
initial and final state for settings entirely through modifications
to the BOTTLEROCKET-DATA filesystem.

Bottlerocket makes relatively few assumptions about that filesystem,
and it's intended to be modified by site-specific workflows such as
caching container images, so this is a natural fit.

Signed-off-by: Ben Cressey <bcressey@amazon.com>