netdog: keep DHCP configuration for interfaces
For managed interfaces using DHCP, set `KeepConfiguration=dhcp` in
the generated network unit. This means that addresses and routes from
the DHCP server will never be dropped even if the lease expires.

This is not allowed by the DHCP specification, but is meant for cases
where the connection should never go away. Arguably that applies to
Bottlerocket since the host will be very difficult to troubleshoot if
the primary interface goes offline.

More importantly for our use case, this prevents the DHCPv4 lease
from being discarded when `netdog` reloads `systemd-networkd` as part
of applying DNS-related settings on boot or after they are modified.
Ideally, the DHCPv6 lease would also be retained, but that isn't
supported by the current implementation of `KeepConfiguration`.

Signed-off-by: Ben Cressey <bcressey@amazon.com>