imdsclient: trim trailing dot from hostname
On Amazon EC2, if a user configures a VPC DHCP Option Set with a domain
name ending in a trailing dot, the local hostname assigned to an
instance as exposed via IMDS will also have a trailing dot. This is not
a valid hostname. Trim the trailing dot the moment we read it from IMDS.

Signed-off-by: Markus Boehme <markubo@amazon.com>