netdog: Add none to bonds for deterministic MACs
networkd will generate a new MAC Address for bonds by default. If
MACAddress is set to none, then it will use the lowest one found which
is closer to the behavior of wicked. Future work can be added to allow a
user to specify the MAC Address used for a bond but currently its hard
coded to use the value of none.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>