netdog: Add BindCarrier to bond network files
networkd will lose track of carrier for worker devices in a bond when
bringing it up. This causes a delay in the bond becoming routable while
the devices re-establish carrier status. This change adds BindCarrier to
bond .network files which ensures networkd brings up the bond if at
least one link has carrier.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>