actions: Do not continue on error
We've set the `fail-fast` setting in the GitHub Action that runs
on every PR to `false`. This means that even if one of the variants
encounters an error, all other variants will continue to run.

This is useful in getting data if a failure only impacts one specific
variant, but in some cases this means that we will run X number of jobs
for all variants that have no chance of succeeding, resulting in a waste
of Action runner resources.

There are times when we hit failures due to intermittent network or
other issues, but there are also times where there are things like
packages being added that have not been put in the lookaside cache. In
either case, someone will need to look at the failure and determine what
went wrong. For those intermittent failures, it's easy enough to trigger
the actions to run again. But if there is a change required to the code
or something that is not just an intermittent failure, this would allow
the necessary actions to take place before wasting any more resources.

This also flips `continue-on-error` from `true` to `false`. Similar to
`fail-fast`, `continue-on-error` allows subsequent steps to run even if
a previous step has failed. While this may provide some useful data, it
can also cause unnecessary processing time for something that will not
pass.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>